refactor: Card 객체에서 initCard 메소드를 Deck 객체로 이동
